binary archive looks like it is working, putting in some static_casts to quiet down some warnings

+++ include/cereal/archives/binary.hpp
@@ -136,14 +136,14 @@ namespace cereal
   template <class T> inline
   void save(BinaryOutputArchive & ar, BinaryData<T> const & bd)
   {
-    ar.saveBinary(bd.data, bd.size);
+    ar.saveBinary(bd.data, static_cast<std::size_t>(bd.size));
   }
 
   //! Loading binary data
   template <class T> inline
   void load(BinaryInputArchive & ar, BinaryData<T> & bd)
   {
-    ar.loadBinary(bd.data, bd.size);
+    ar.loadBinary(bd.data, static_cast<std::size_t>(bd.size));
   }
 } // namespace cereal
